refactor: simplify MCP connection UI and remove unused components
- Updated translations for MCP server URL across multiple languages.
- Removed unnecessary components related to MCP setup instructions and states.
- Streamlined the MCP connection handling in the UI for better clarity and usability.
