+++ apps/web/messages/en.json
@@ -2456,33 +2456,10 @@
       }
     },
     "MCP": {
-      "generateButton": "Show Sokosumi MCP URL",
-      "urlLabel": "Sokosumi MCP Server URL",
+      "urlLabel": "MCP Server URL",
       "existingKeyNote": "This hosted MCP connection uses OAuth. Your MCP client will ask you to sign in with Sokosumi when connecting.",
-      "existingKeyNote2": "No API key needs to be copied into the client.",
-      "existingKey": "YOUR_EXISTING_KEY",
-      "keyDisabled": "Your MCP connection exists but is disabled.",
-      "keyDisabledNote": "You can enable it or create a new one.",
-      "enableKey": "Enable Key",
-      "createNewKey": "Create New Key",
-      "regenerateKey": "Refresh URL",
-      "setupPreviewTitle": "How to connect with OAuth",
       "docsLink": "View MCP Documentation",
-      "loadError": "Failed to load MCP connection information. Please try again.",
-      "copyError": "Failed to copy URL to clipboard",
-      "retry": "Retry",
-      "loading": "Loading MCP connection information...",
-      "Instructions": {
-        "title": "How to connect an MCP client to Sokosumi",
-        "step1": "Open your MCP client settings",
-        "step2": "Click on \"Connectors\"",
-        "step3": "Click on \"Add custom connector\"",
-        "step4": "Enter a name (e.g., \"Sokosumi\") and paste the MCP server URL in the URL field. Leave Advanced Settings empty. Click \"Add\"",
-        "step5": {
-          "text": "Complete the Sokosumi sign-in when your client opens the browser. Then try asking:",
-          "example": "What agents are currently available on Sokosumi?"
-        }
-      }
+      "copyError": "Failed to copy URL to clipboard"
     },
     "Account": {
       "Name": {
