+++ apps/web/messages/fr.json
@@ -2456,33 +2456,10 @@
       }
     },
     "MCP": {
-      "generateButton": "Afficher l’URL MCP Sokosumi",
-      "urlLabel": "URL du serveur MCP Sokosumi",
+      "urlLabel": "URL du serveur MCP",
       "existingKeyNote": "Cette connexion MCP hébergée utilise OAuth. Votre client MCP vous demandera de vous connecter à Sokosumi lors de la connexion.",
-      "existingKeyNote2": "Aucune clé API ne doit être copiée dans le client.",
-      "existingKey": "YOUR_EXISTING_KEY",
-      "keyDisabled": "Votre connexion MCP existe mais est désactivée.",
-      "keyDisabledNote": "Vous pouvez l’activer ou en créer une nouvelle.",
-      "enableKey": "Activer la clé",
-      "createNewKey": "Créer une nouvelle clé",
-      "regenerateKey": "Actualiser l’URL",
-      "setupPreviewTitle": "Connexion avec OAuth",
       "docsLink": "Voir la documentation MCP",
-      "loadError": "Impossible de charger les informations de connexion MCP. Veuillez réessayer.",
-      "copyError": "Échec de la copie de l’URL dans le presse-papiers",
-      "retry": "Réessayer",
-      "loading": "Chargement des informations de connexion MCP...",
-      "Instructions": {
-        "title": "Comment connecter un client MCP à Sokosumi",
-        "step1": "Ouvrez les paramètres de votre client MCP",
-        "step2": "Cliquez sur \"Connectors\"",
-        "step3": "Cliquez sur \"Add custom connector\"",
-        "step4": "Saisissez un nom (par ex. \"Sokosumi\") et collez l’URL du serveur MCP dans le champ URL. Laissez les paramètres avancés vides. Cliquez sur \"Add\"",
-        "step5": {
-          "text": "Terminez la connexion à Sokosumi lorsque votre client ouvre le navigateur. Essayez ensuite de demander :",
-          "example": "Quels agents sont actuellement disponibles sur Sokosumi ?"
-        }
-      }
+      "copyError": "Échec de la copie de l’URL dans le presse-papiers"
     },
     "Account": {
       "Name": {
