+++ apps/web/src/app/(app)/components/MCP/mcp-url.ts
@@ -1,0 +1,6 @@
+import { getEnvPublicConfig } from "@/config/env.public";
+
+export function getMcpUrl(): string {
+  const baseUrl = getEnvPublicConfig().NEXT_PUBLIC_MCP_URL.replace(/\/$/, "");
+  return `${baseUrl}/mcp`;
+}
